.history/backend/requirements_20251003231028.txt
alembic==1.14.0
blinker==1.9.0
click==8.3.0
Flask==3.1.2
flask-bcrypt==1.0.1
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
PyJWT==2.10.1
pysqlcipher3==1.2.0
python-dotenv==1.1.1
SQLAlchemy==2.0.43
typing_extensions==4.15.0
